% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiomics.R
\name{addRepsToMapping}
\alias{addRepsToMapping}
\title{addRepsToMapping}
\usage{
addRepsToMapping(curr_mapping, curr_variable)
}
\arguments{
\item{data_frame, }{string}
}
\value{
The mapping data frame specified in input, but with replicate values integrated
}
\description{
Make sure that data frame does not contain any illegal values.
}
\examples{
mapping_with_reps <- addRepsToMapping(mapping, a_variable_name);
}
