% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiomics.R
\name{addRepsToMappingTwoVariables}
\alias{addRepsToMappingTwoVariables}
\title{addRepsToMappingTwoVariables}
\usage{
addRepsToMappingTwoVariables(curr_mapping, facet1, facet2)
}
\value{
Nothing. If data frame is offensive, will call stop()
}
\description{
Make sure that data frame does not contain any illegal values.
}
\examples{
mapping <- validate_data_frame(mapping);
}
