% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{filter_dist_outliers}
\alias{filter_dist_outliers}
\title{Filter distance matrix for outliers}
\usage{
filter_dist_outliers(d, quant = 0.8, n_neighbour = 1)
}
\arguments{
\item{d}{distance matrix as matrix}

\item{quant}{quantile of all distances to use as a filtering threshold}

\item{n_neighbour}{number of neighbours required}
}
\value{
filtered distance matrix, or NULL if all samples were pruned
}
\description{
Detect and filter distance matrix for potential outliers by heuristical detection.
The distance matrix is pruned using the following criterion until no further outliers
are found:
\enumerate{
\item Compute distance threshold by computing \code{quant} quantile of the entire distance matrix
\item Find and remove the most extreme data point defined by highest median distance to other
data if it has less than \code{n_neighbour} neighbours closer than threshold to itself
\item  Return to the first step and continue until no further data points are pruned.
}
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}

