% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{gene_distance}
\alias{gene_distance}
\title{Compute gene similarity (or distance) for gene profiles}
\usage{
gene_distance(x, return_similarity = FALSE)
}
\arguments{
\item{x}{data frame or matrix (genomes x genes) containing gene profiles}

\item{return_similarity}{set TRUE to return similarity instead of distance}
}
\value{
a dist object containing gene similarities / distances
}
\description{
This functions returns gene similarity / distance based on microbial gene profiles 
given as input. Gene similarity between two clades / genomes is the number of common
genes between the genomes divided by the number of genes in the smaller genome,
as defined in https://www.ncbi.nlm.nih.gov/pubmed/9916801
Gene distance is '1 - gene similarity'.
#'
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}
