% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{get_significant_values}
\alias{get_significant_values}
\title{A function to select significant components among matrix of measurements}
\usage{
get_significant_values(values, p_values, q_values = NA,
  pvalue_threshold = 0.01, qvalue_threshold = 0.1)
}
\arguments{
\item{values}{matrix of values of interest (e.g. correlations)}

\item{p_values}{matrix of p-values for the matrix in the first argument}

\item{q_values}{(optional) matrix of q-values for the matrix in the first argument. If q-values are given, no p-value thresholding is done.}

\item{pvalue_threshold}{threshold for significance for the p-values (default 0.01)}

\item{qvalue_threshold}{threshold for significance for the q-value (default 0.1)}
}
\value{
a list with following elements
\itemize{
    \item values: matrix of significant values in the matrix of the first argument 
    \item p: matrix of p-values for the values above
    \item q: matrix of q-values for the values above (if q-values are given as argument)
}
}
\description{
This function takes a matrix of values (e.g. correlations from \code{\link{pairwise_spearman}})
and corresponding p-values and/or q-values. Given the threshold for p- and q-values 
it will return reduced matrices of significant values where each row and column 
contain at least one significant value (while retaining the original row and 
column names).
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}
