% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{jaccard}
\alias{jaccard}
\title{Compute Jaccard distance / index for given microbial profiles}
\usage{
jaccard(x, return_index = FALSE)
}
\arguments{
\item{x}{data frame or matrix containing microbial community profiles}

\item{return_index}{set TRUE to return Jaccard Index instead of distance}
}
\value{
a dist object containing Jaccard distances / indeces
}
\description{
This functions computes distance / similarity matrix based on microbial community profiles
given as input. Jaccard distance is based on Jaccard index, following the definition given
in wikipedia (https://en.wikipedia.org/wiki/Jaccard_index). Note, that this is different
from Jaccard index in vegdist function in package vegan.
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}

