% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{Maaslin.wrapper}
\alias{Maaslin.wrapper}
\title{A wrapper function for MaAsLin}
\usage{
Maaslin.wrapper(taxa, metadata, strOutputDIR = tempfile(),
  variables = colnames(metadata), strRandomCovariates = NULL, ...)
}
\arguments{
\item{taxa}{input taxa (samples x taxa data.frame)}

\item{metadata}{input metadata (samples x features data.frame)}

\item{strOutputDIR}{output directory (defaults to tmp directory)}

\item{variables}{subset of metadata columns to test (defaults to all columns)}

\item{strRandomCovariates}{features to be used as random effects}

\item{\dots}{additional parameters for the Maaslin call. Parameters passed to
\code{\link[Maaslin]{Maaslin}}.}
}
\value{
a list with an association table per feature with associations with taxa.
}
\description{
This function will run MaAsLin (\url{https://bitbucket.org/biobakery/maaslin/}) given a
set of taxa and metadata.
}
\details{
This function requires Maaslin-package (\url{https://bitbucket.org/biobakery/maaslin/}).
However, in order to avoid multiple dependencies this is not explicitely defined as
dependency in the microbiomics-package.
}
\author{
Tommi Vatanen <tommivat@gmail.com>, Raivo Kolde
}

