% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{Maaslin2.wrapper}
\alias{Maaslin2.wrapper}
\title{A wrapper function for MaAsLin2}
\usage{
Maaslin2.wrapper(taxa, metadata, output = tempfile(), fixed_effects,
  random_effects = NULL, ...)
}
\arguments{
\item{taxa}{input taxa (samples x taxa data.frame)}

\item{metadata}{input metadata (samples x features data.frame)}

\item{output}{output directory (defaults to tmp directory)}

\item{fixed_effects}{subset of metadata columns to test (defaults to all columns)}

\item{random_effects}{features to be used as random effects}

\item{\dots}{additional parameters for the Maaslin call. Parameters passed to 
\code{\link[Maaslin2]{Maaslin}}.}
}
\value{
a list with an association table per feature with associations with taxa.
}
\description{
This function will run MaAsLin2 (\url{https://bitbucket.org/biobakery/maaslin2/}) given a 
set of taxa and metadata.
 
This function requires Maaslin2-package (\url{https://bitbucket.org/biobakery/maaslin2/}).
However, in order to avoid multiple dependencies this is not explicitely defined as
dependency in the microbiomics-package.
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}
