% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{pairwise_spearman}
\alias{pairwise_spearman}
\title{A function to compute pairwise correlations between two data sets}
\usage{
pairwise_spearman(x, y)
}
\arguments{
\item{x}{data set 1 (samples x features1)}

\item{y}{data set 2 (samples x features2)}
}
\value{
a list with following elements
\itemize{
    \item r: features1 x features2 matrix of pairwise correlations
    \item p: features1 x features2 matrix of nominal p-values
    \item q: features1 x features2 matrix of fdr corrected p-values (q-values)
}
}
\description{
This function computes pairwise correlations between columns of two data sets.
It assumes that the data sets have same samples (on rows) in same order.
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}
