% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiomics.R
\name{processTaxonomicLineages}
\alias{processTaxonomicLineages}
\title{processTaxonomicLineages}
\usage{
processTaxonomicLineages(tTaxonomy3, summarize_lineage)
}
\value{
data_frame
}
\description{
Make sure that data frame does not contain any illegal values.
}
\examples{
data_frame_with_processed_taxa_lineages <- processTaxonomicLineages(data_frame, summarize_lineages=TRUE/FALSE);
}
