% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{read_metaphlan_table}
\alias{read_metaphlan_table}
\title{A function to read MetaPhlAn 2.0 output file}
\usage{
read_metaphlan_table(filename, kingdom = "k__Bacteria", lvl = 7)
}
\arguments{
\item{filename}{MetaPhlAn output file to be read}

\item{kingdom}{to read: One of "k__Bacteria", "k__Viruses", "k__Eukaryota"}

\item{lvl}{what level of taxonomy to read (default = 7, species)}
}
\value{
a data frame with the MetaPhlAn results
}
\description{
This functions read a MetaPhlAn 2.0 output file as an R data frame
}
\author{
Tommi Vatanen tommivat@gmail.com
}

