% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomics.r
\name{read_taxon_table}
\alias{read_taxon_table}
\title{A function to read taxon table produced by summarize_taxa.py}
\usage{
read_taxon_table(filename, lvl = 6)
}
\arguments{
\item{filename}{taxon table file to read}

\item{lvl}{taxonomic level to read (default = 6, genera)}
}
\value{
a data frame with the taxonomic data
}
\description{
This functions reads a summarize_taxa.py output file as an R data frame
}
\author{
Tommi Vatanen <tommivat@gmail.com>
}
