% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiomics-package.r
\docType{package}
\name{microbiomics-package}
\alias{microbiomics-package}
\title{Microbiomics}
\description{
Implementation of some functions that maybe handy
in analysing microbiome data
}

