% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiomics-package.r
\docType{package}
\name{microbiomics-package}
\alias{microbiomics-package}
\title{Microbiomics}
\description{
Implementation of some functions that may come handy in analysing microbiome data.
}
\details{
Following functions are included:
\describe{
 \item{\code{\link{read_metaphlan_table}}}{read a MetaPhlAn 2.0 output file as an R data frame}
 \item{\code{\link{read_taxon_table}}}{read a summarize_taxa.py output file as an R data frame}
 \item{ \code{\link{pairwise_spearman}}}{compute pairwise correlations between columns of two data sets}
 \item{\code{\link{get_significant_values}}}{get significant components among a matrix of measurements}
 \item{\code{\link{Maaslin.wrapper}}}{run MaAsLin (\url{https://bitbucket.org/biobakery/maaslin/}) given a
set of taxa and metadata}
}
}

